исправлены замечания
